+++ app/models/auth_provider.rb
@@ -11,16 +11,17 @@ class AuthProvider < ActiveRecord::Base
     end
   end
 
-  PROVIDABLE_TYPES = %w(DatabaseProvider OAuth2Provider)
+  PROVIDABLE_TYPES = %w[DatabaseProvider OAuth2Provider].freeze
 
-  belongs_to :providable, :polymorphic => true, dependent: :destroy
+  belongs_to :providable, polymorphic: true, dependent: :destroy
   accepts_nested_attributes_for :providable
   attr_accessible :name, :providable_type, :providable_attributes
 
   before_create :set_initial_state
 
-  def build_providable(params, assignment_options)
+  def build_providable(params)
     raise "Unknown providable_type: #{providable_type}" unless PROVIDABLE_TYPES.include?(providable_type)
+
     self.providable = providable_type.constantize.new(params)
   end
 
@@ -30,11 +31,9 @@ class AuthProvider < ActiveRecord::Base
 
     begin
       provider = find_by(status: 'active')
-      if provider.nil?
-        return local
-      else
-        return provider
-      end
+      return local if provider.nil?
+
+      return provider
     rescue ActiveRecord::StatementInvalid
       # we fall here on database creation because the table "active_providers" still does not exists at the moment
       return local
@@ -43,13 +42,12 @@ class AuthProvider < ActiveRecord::Base
 
   ## Get the provider matching the omniAuth strategy name
   def self.from_strategy_name(strategy_name)
-    if strategy_name.blank? or all.empty?
-      return SimpleAuthProvider.new
-    end
+    return SimpleAuthProvider.new if strategy_name.blank? || all.empty?
+
     parsed = /^([^-]+)-(.+)$/.match(strategy_name)
     ret = nil
     all.each do |strategy|
-      if strategy.provider_type == parsed[1] and strategy.name.downcase.parameterize == parsed[2]
+      if strategy.provider_type == parsed[1] && strategy.name.downcase.parameterize == parsed[2]
         ret = strategy
         break
       end
@@ -59,7 +57,7 @@ class AuthProvider < ActiveRecord::Base
 
   ## Return the name that should be registered in OmniAuth for the corresponding strategy
   def strategy_name
-    provider_type+'-'+name.downcase.parameterize
+    provider_type + '-' + name.downcase.parameterize
   end
 
   ## Return the provider type name without the "Provider" part.
@@ -81,7 +79,7 @@ class AuthProvider < ActiveRecord::Base
   end
 
   def safe_destroy
-    if self.status != 'active'
+    if status != 'active'
       destroy
     else
       false
@@ -89,9 +87,10 @@ class AuthProvider < ActiveRecord::Base
   end
 
   private
+
   def set_initial_state
     # the initial state of a new AuthProvider will be 'pending', except if there is currently
     # no providers in the database, he we will be 'active' (see seeds.rb)
-    self.status = 'pending' unless AuthProvider.count == 0
+    self.status = 'pending' unless AuthProvider.count.zero?
   end
 end