+++ scripts/elastic-upgrade.sh
@@ -36,6 +36,12 @@ config()
     echo "sudo was not found, exiting..."
     exit 1
   fi
+  if ! command -v awk || ! [[ $(awk -W version) =~ ^GNU ]]
+  then
+    echo "Please install GNU Awk before running this script."
+    echo "gawk was not found, exiting..."
+    exit 1
+  fi
   echo "checking memory..."
   mem=$(free -mt | grep Total | awk '{print $2}')
   if [ "$mem" -lt 4000 ]